feat: add clip similarity calculation. add RAM++ to as a small powerful model to list objects.
Update dependencies, modify configuration settings, and implement RAMLister for object detection. The pyproject.toml was updated to include new packages such as decord, ftfy, einops, timm, and wandb. The config.toml was adjusted for model settings, and the processing manager was enhanced to support feature extraction. Additionally, the README was updated for clarity.